# api/oas-doc.yaml
openapi: 3.0.0
info:
  title: School Data Api API
  version: '1.0'
  description: 'This API serves students info, courses info and grades info.'
  license:
    name: ''
  contact:
    name: kihun han
    url: kihun.net
    email: kihun@kihun.net
paths:
...
  '/students/:Id/addCourses':
    post:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Name:
                    type: string
                  Class:
                    type: string
                  enrolledCourses:
                    type: array
                    items:
                      type: integer
                  Address:
                    type: string
              examples:
                response:
                  value:
                    Id: et tempor quis Ut consectetur
                    Name: dolore ma
                    Class: esse in reprehenderit p
                    enrolledCourses:
                      - 82622458
                      - 93544276
                    Address: velit in consequat
      description: This path will enroll the courses to the student data
      summary: Post Courses by Student Id
      operationId: postCoursesByStudentId
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Student Id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Course:
                  type: integer
        x-examples:
          application/json:
            Course: 16076744
  '/students/:Id/courses':
    get:
      responses:
        '200':
          description: ''
          content:
            text/html:
              schema:
                type: object
                properties: {}
              examples:
                response:
                  value: 'Enrolled courses are course1, course2, course3'
      summary: Get enrolled Courses
      description: This path will return specific students enrolled courses
      operationId: getEnrolledCourses
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Student Id
          schema:
            type: integer
...
servers:
  - url: 'http://localhost:3000/'